LayoutTests/js/get-by-pname-expected.txt

This tests that op_get_by_pname is compiled correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test tier: cold
PASS foo1(o) is 11
PASS foo2(p) is 10
PASS foo3(q) is 3467
PASS foo4(r) is 113
PASS foo5(s) is 182
PASS getByPnameOnConstant1(a) is 0
PASS getByPnameOnVar11(a, 100) is 0
PASS getByPnameOnVar12(a, 'abc') is '0abc'
PASS getByPnameOnVar13(a, o) is 0
PASS getByPnameOnVar14(a, o1) is 4
PASS getByPnameOnVar15(a, a) is 6
PASS getByPnameOnConstant2(o1) is 0
PASS getByPnameOnVar21(o1, 100) is 0
PASS getByPnameOnVar22(o1, 'abc') is '0bc0'
PASS getByPnameOnVar23(o1, o) is 0
PASS getByPnameOnVar24(o1, o1) is 11
PASS getByPnameOnVar25(o1, a) is 5
PASS getByPnameOnConstant3(o) is 0
PASS getByPnameOnVar31(o, 100) is 0
PASS getByPnameOnVar32(o, 'abc') is 0
PASS getByPnameOnVar33(o, o) is 11
PASS getByPnameOnVar34(o, o1) is 0
PASS getByPnameOnVar35(o, a) is 0
PASS getByPnameOnConstant4(0) is 0
PASS getByPnameOnVar41(0, 100) is 0
PASS getByPnameOnVar42(0, 'abc') is 0
PASS getByPnameOnVar43(0, o) is 0
PASS getByPnameOnVar44(0, o1) is 0
PASS getByPnameOnVar45(0, a) is 0
PASS getByPnameOnConstant5('abc') is 0
PASS getByPnameOnVar51('abc', 100) is 0
PASS getByPnameOnVar52('abc', 'abc') is '0abc'
PASS getByPnameOnVar53('abc', o) is 0
PASS getByPnameOnVar54('abc', o1) is 4
PASS getByPnameOnVar55('abc', a) is 6
PASS getByPnameOnVar56('def', 'abc') is '0abc'

Test tier: llint
PASS foo1(o) is 11
PASS foo2(p) is 10
PASS foo3(q) is 3467
PASS foo4(r) is 113
PASS foo5(s) is 182
PASS getByPnameOnConstant1(a) is 0
PASS getByPnameOnVar11(a, 100) is 0
PASS getByPnameOnVar12(a, 'abc') is '0abc'
PASS getByPnameOnVar13(a, o) is 0
PASS getByPnameOnVar14(a, o1) is 4
PASS getByPnameOnVar15(a, a) is 6
PASS getByPnameOnConstant2(o1) is 0
PASS getByPnameOnVar21(o1, 100) is 0
PASS getByPnameOnVar22(o1, 'abc') is '0bc0'
PASS getByPnameOnVar23(o1, o) is 0
PASS getByPnameOnVar24(o1, o1) is 11
PASS getByPnameOnVar25(o1, a) is 5
PASS getByPnameOnConstant3(o) is 0
PASS getByPnameOnVar31(o, 100) is 0
PASS getByPnameOnVar32(o, 'abc') is 0
PASS getByPnameOnVar33(o, o) is 11
PASS getByPnameOnVar34(o, o1) is 0
PASS getByPnameOnVar35(o, a) is 0
PASS getByPnameOnConstant4(0) is 0
PASS getByPnameOnVar41(0, 100) is 0
PASS getByPnameOnVar42(0, 'abc') is 0
PASS getByPnameOnVar43(0, o) is 0
PASS getByPnameOnVar44(0, o1) is 0
PASS getByPnameOnVar45(0, a) is 0
PASS getByPnameOnConstant5('abc') is 0
PASS getByPnameOnVar51('abc', 100) is 0
PASS getByPnameOnVar52('abc', 'abc') is '0abc'
PASS getByPnameOnVar53('abc', o) is 0
PASS getByPnameOnVar54('abc', o1) is 4
PASS getByPnameOnVar55('abc', a) is 6
PASS getByPnameOnVar56('def', 'abc') is '0abc'

Test tier: baseline
PASS foo1(o) is 11
PASS foo2(p) is 10
PASS foo3(q) is 3467
PASS foo4(r) is 113
PASS foo5(s) is 182
PASS getByPnameOnConstant1(a) is 0
PASS getByPnameOnVar11(a, 100) is 0
PASS getByPnameOnVar12(a, 'abc') is '0abc'
PASS getByPnameOnVar13(a, o) is 0
PASS getByPnameOnVar14(a, o1) is 4
PASS getByPnameOnVar15(a, a) is 6
PASS getByPnameOnConstant2(o1) is 0
PASS getByPnameOnVar21(o1, 100) is 0
PASS getByPnameOnVar22(o1, 'abc') is '0bc0'
PASS getByPnameOnVar23(o1, o) is 0
PASS getByPnameOnVar24(o1, o1) is 11
PASS getByPnameOnVar25(o1, a) is 5
PASS getByPnameOnConstant3(o) is 0
PASS getByPnameOnVar31(o, 100) is 0
PASS getByPnameOnVar32(o, 'abc') is 0
PASS getByPnameOnVar33(o, o) is 11
PASS getByPnameOnVar34(o, o1) is 0
PASS getByPnameOnVar35(o, a) is 0
PASS getByPnameOnConstant4(0) is 0
PASS getByPnameOnVar41(0, 100) is 0
PASS getByPnameOnVar42(0, 'abc') is 0
PASS getByPnameOnVar43(0, o) is 0
PASS getByPnameOnVar44(0, o1) is 0
PASS getByPnameOnVar45(0, a) is 0
PASS getByPnameOnConstant5('abc') is 0
PASS getByPnameOnVar51('abc', 100) is 0
PASS getByPnameOnVar52('abc', 'abc') is '0abc'
PASS getByPnameOnVar53('abc', o) is 0
PASS getByPnameOnVar54('abc', o1) is 4
PASS getByPnameOnVar55('abc', a) is 6
PASS getByPnameOnVar56('def', 'abc') is '0abc'

Test tier: dfg
PASS foo1(o) is 11
PASS foo2(p) is 10
PASS foo3(q) is 3467
PASS foo4(r) is 113
PASS foo5(s) is 182
PASS getByPnameOnConstant1(a) is 0
PASS getByPnameOnVar11(a, 100) is 0
PASS getByPnameOnVar12(a, 'abc') is '0abc'
PASS getByPnameOnVar13(a, o) is 0
PASS getByPnameOnVar14(a, o1) is 4
PASS getByPnameOnVar15(a, a) is 6
PASS getByPnameOnConstant2(o1) is 0
PASS getByPnameOnVar21(o1, 100) is 0
PASS getByPnameOnVar22(o1, 'abc') is '0bc0'
PASS getByPnameOnVar23(o1, o) is 0
PASS getByPnameOnVar24(o1, o1) is 11
PASS getByPnameOnVar25(o1, a) is 5
PASS getByPnameOnConstant3(o) is 0
PASS getByPnameOnVar31(o, 100) is 0
PASS getByPnameOnVar32(o, 'abc') is 0
PASS getByPnameOnVar33(o, o) is 11
PASS getByPnameOnVar34(o, o1) is 0
PASS getByPnameOnVar35(o, a) is 0
PASS getByPnameOnConstant4(0) is 0
PASS getByPnameOnVar41(0, 100) is 0
PASS getByPnameOnVar42(0, 'abc') is 0
PASS getByPnameOnVar43(0, o) is 0
PASS getByPnameOnVar44(0, o1) is 0
PASS getByPnameOnVar45(0, a) is 0
PASS getByPnameOnConstant5('abc') is 0
PASS getByPnameOnVar51('abc', 100) is 0
PASS getByPnameOnVar52('abc', 'abc') is '0abc'
PASS getByPnameOnVar53('abc', o) is 0
PASS getByPnameOnVar54('abc', o1) is 4
PASS getByPnameOnVar55('abc', a) is 6
PASS getByPnameOnVar56('def', 'abc') is '0abc'

PASS successfullyParsed is true

TEST COMPLETE
